Fix test script and add to build pipeline
