+++ src/VSCodeExtension/package.json.v.template
@@ -123,14 +123,15 @@
         "vscode": "^1.1.34",
         "typescript": "^2.6.1",
         "tslint": "^5.8.0",
-        "mocha": "^5.2.0",
+        "ts-node": "^8.10.2",
+        "mocha": "^8.0.0",
         "@types/yeoman-generator": "^3.1.4",
         "@types/yosay": "^0.0.29",
         "yeoman-test": "^1.7.0",
         "yeoman-assert": "^3.1.0",
         "@types/yeoman-environment": "^2.3.3",
         "@types/node": "^9.6.5",
-        "@types/mocha": "^5.2.0",
+        "@types/mocha": "^8.0.0",
         "@types/which": "1.3.1",
         "@types/tmp": "0.0.33",
         "@types/semver": "^6.0.0",
